// frontend/src/generated/api/models/FurrowCellInfo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CellInfo } from './CellInfo';
export type FurrowCellInfo = {
    furrowWidth: number;
    furrowLength: number;
    minPlantSpacing: number;
    rows: number;
    CellInfoArray: Array<CellInfo>;
};